+++ app/src/routes/index.ts
@@ -1,5 +1,7 @@
 import { RouteObject } from "react-router-dom";
-import { rulesRoutes } from "./rules.route";
-import { sessionsRoutes } from "./sessions.route";
+import { ruleRoutes } from "./rules-route";
+import { sessionRoutes } from "./sessions-route";
+import { apiClientRoutes } from "./apiClient-route";
+import { accountRoutes } from "./account-route";
 
-export const updatedRoutes: RouteObject[] = [...rulesRoutes, ...sessionsRoutes];
+export const updatedRoutes: RouteObject[] = [...ruleRoutes, ...sessionRoutes, ...apiClientRoutes, ...accountRoutes];
